requirements.txt
amqp==5.1.1 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_full_version <= "3.11.2"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
billiard==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
celery==5.3.4 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
django-cors-headers==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
django-extensions==3.2.3 ; python_version >= "3.11" and python_version < "4.0"
django-filter==23.3 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.5 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework-simplejwt==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
drf-spectacular==0.26.4 ; python_version >= "3.11" and python_version < "4.0"
eventlet==0.33.3 ; python_version >= "3.11" and python_version < "4.0"
feedparser==6.0.10 ; python_version >= "3.11" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2023.7.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.19.1 ; python_version >= "3.11" and python_version < "4.0"
kombu==5.3.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.11" and python_version < "4.0"
prompt-toolkit==3.0.39 ; python_version >= "3.11" and python_version < "4.0"
psycopg==3.1.10 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
redis==4.6.0 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.30.2 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.10.3 ; python_version >= "3.11" and python_version < "4.0"
sgmllib3k==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.5 ; python_version >= "3.11" and python_version < "4.0"
vine==5.0.0 ; python_version >= "3.11" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.11" and python_version < "4.0"

Django~=4.2.5
celery~=5.3.4
python-dotenv~=1.0.0
djangorestframework~=3.14.0
feedparser~=6.0.10
PyJWT~=2.8.0